Refactored `Material`
