+++ analysis_options.yaml
@@ -70,7 +70,7 @@ linter:
   - avoid_slow_async_io
   - avoid_type_to_string
   - avoid_types_as_parameter_names
-  # - avoid_types_on_closure_parameters
+  - avoid_types_on_closure_parameters
   - avoid_unnecessary_containers
   - avoid_unused_constructor_parameters
   - avoid_void_async
