+++ lib/playx_3d_scene.dart
@@ -13,8 +13,6 @@ export 'src/models/scene/indirect_light/indirect_light.dart';
 export 'src/models/scene/indirect_light/ktx_indirect_light.dart';
 export 'src/models/scene/light/light.dart';
 export 'src/models/scene/material/material.dart';
-export 'src/models/scene/material/material_parameter.dart';
-export 'src/models/scene/material/material_type.dart';
 export 'src/models/scene/material/texture/texture.dart';
 export 'src/models/scene/material/texture/texture_sampler.dart';
 export 'src/models/scene/material/texture/enums/mag_filter.dart';
